Subsonic: Move to new api library
I refactored the open subsonic api library to better match the
specification documentation and to ensure that the objects we return
meet that documentation. The new libary also adds typing hints and
removed a bunch of hand rolled parsing in favor of mashumaro classes.

This is a big change and I believe I have tested every possible
interaction with the provider (and updated the tests) so that it should
all be working.

Signed-off-by: Eric B Munson <eric@munsonfam.org>