chore/fix: Subsonic: Update parsers and tests
We have metadata associated with tracks, albums, and artists that we are
not passing on to MA. There isn't a reason to hold anything back so make
sure we populate any and all fields that we can when we parse to MA
objects.

Take this opportunity to move the last parser out of the provider class
and add a test for it.

Signed-off-by: Eric B Munson <eric@munsonfam.org>