Feat: Subsonic: Add configurable recommendations
The provider has the ability to ask for things like newest albums, most
played albums, and anything that is starred by a user. These could be
used to fill out the recommendation rows in the home page if the user
wants. Add configuration for what kinds (if any) of these
recommendations are desired and the size of the list.

Signed-off-by: Eric B Munson <eric@munsonfam.org>